gf3208: Always use USE_PLATFORM_BUS
The USE_PLATFORM_BUS macro definition is enabled by default,
so assume to always use USE_PLATFORM_BUS and then remove
the macro definitions and checks. This also removes the
code in the USE_SPI_BUS macro definition check, because
USE_SPI_BUS is not used so just delete it.

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>